getting food

+++ arduino_code/ArduinoProject/servo.h
@@ -11,7 +11,7 @@
 namespace servo 
 {
   void init();
-  void set_positions(float amplitude, float wavelength, unsigned long time_milli, int wavetype, int side);
+  void set_positions(float amplitude, float wavelength, float time_milli, int wavetype, int side);
   void set_neutrals(int side);
 }
